+++ app/src/main/java/com/vaibhav/healthify/data/models/remote/UserDTO.kt
@@ -4,7 +4,7 @@ data class UserDTO(
     val username: String = "",
     val email: String = "",
     val profileImg: String = "",
-    val exp: Int = 0,
+    val exp: Long = 0L,
     val waterLimit: Int = 0,
     val sleepLimit: Int = 0,
     val age: Int = 0,
